pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prism-ai-router"
version = "0.1.0"
description = "Multi-Modal AI Router & Orchestrator - Intelligent routing for text, image, audio, video, and document processing"
readme = "README.md"
license = "MIT"
authors = [
    { name = "moggan1337", email = "moggan1337@users.noreply.github.com" }
]
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9.0",
    "redis>=5.0.0",
    "pydantic>=2.5.0",
    "structlog>=24.0.0",
    "tenacity>=8.2.0",
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-exporter-otlp>=1.21.0",
    "pillow>=10.0.0",
    "python-magic>=0.4.27",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "DTZ", "T10", "ISC", "ICN", "PIE", "PYI", "RSE", "RET", "SLF", "SIM", "TID", "ARG", "PLE", "PLR", "PLW"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
